+++ cmd/deploy.go
@@ -1,0 +1,34 @@
+/*
+Copyright © 2025 Chanhyeok Seo chanhyeok.seo2@gmail.com
+*/
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+// deployCmd represents the deploy command
+var deployCmd = &cobra.Command{
+	Use:   "deploy",
+	Short: "Deploy Dumie manager to manage instances",
+	Long:  `TODO`,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("deploy called")
+	},
+}
+
+func init() {
+	rootCmd.AddCommand(deployCmd)
+
+	// Here you will define your flags and configuration settings.
+
+	// Cobra supports Persistent Flags which will work for this command
+	// and all subcommands, e.g.:
+	// deployCmd.PersistentFlags().String("foo", "", "A help for foo")
+
+	// Cobra supports local flags which will only run when this command
+	// is called directly, e.g.:
+	// deployCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+}
