--- dumie-cli/main.go
@@ -1,10 +1,0 @@
-/*
-Copyright © 2025 Chanhyeok Seo chanhyeok.seo2@gmail.com
-*/
-package main
-
-import "github.com/dumie-org/dumie-cli/cmd"
-
-func main() {
-	cmd.Execute()
-}
